Cache remote imports in-memory.
This will cache all remote imports as long as the process is
running (frozen imports as well as unfrozen ones). Leads to a big
speedup even for frozen files, because parsing the files from disk
takes considerable times when their normalized forms are large (e.g.
dhall-to-cabal is half a megabyte and takes 2–3 seconds to parse).
This made in-editor checks unbearably slow, even if a dependency
wasn’t actually used.

There might still be speedups possible which we haven’t considered.